requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.33
psycopg2==2.9.5
requests==2.28.2
schedule==1.1.0
boto3==1.26.13
razorpay==1.2.1
pytest==7.4.0
unittest2==1.1.0
docker==5.0.3
flask_wtf==1.0.1
APScheduler==3.9.1
Flask-Cors==3.1.1
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
